# backend/api-spec-complete.yaml
openapi: 3.0.3
info:
  title: RetailGenie API
  description: |
    **RetailGenie** - Comprehensive retail management API with AI-powered recommendations,
    inventory management, user authentication, and real-time analytics.

    ## 🚀 Features
    - 🛍️ **Product Management**: Full CRUD operations for product catalog
    - 🔐 **Authentication**: Secure user registration and login system
    - 🤖 **AI Assistant**: Intelligent product recommendations and chat
    - 📊 **Analytics**: Real-time business insights and dashboards
    - 💰 **Pricing**: Dynamic pricing optimization and competitor analysis
    - 📦 **Inventory**: Stock tracking and management
    - 💬 **Feedback**: Customer reviews and rating system
    - 🔌 **WebSocket**: Real-time notifications and updates
    - 🔄 **Background Tasks**: Asynchronous processing with Celery

    ## 🔐 Authentication
    Most endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## 📈 API Versioning
    This API supports multiple versions:
    - **v1**: Stable production API (current)
    - **v2**: Enhanced API with advanced features

    Use version-specific endpoints: `/api/v1/...` or `/api/v2/...`

    ## ⚡ Rate Limiting
    API requests are limited to:
    - **1000 requests/hour** for authenticated users
    - **100 requests/hour** for unauthenticated users

    ## 📄 Pagination
    List endpoints support pagination with standard parameters:
    - `page`: Page number (default: 1)
...

...

...

...
servers:
...
paths:
...
  /api/v1/products:
    get:
      summary: Get all products (v1)
      description: Retrieve a paginated list of all products with optional filtering
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/InStockFilter'
        - $ref: '#/components/parameters/MinPriceFilter'
        - $ref: '#/components/parameters/MaxPriceFilter'
      responses:
        '200':
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new product (v1)
      description: Create a new product in the catalog
      tags:
        - Products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/products/{product_id}:
    get:
      summary: Get product by ID (v1)
      description: Retrieve a specific product by its ID
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update product (v1)
      description: Update an existing product
      tags:
        - Products
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete product (v1)
      description: Delete a product from the catalog
      tags:
        - Products
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Enhanced V2 Product Endpoints
  /api/v2/products:
    get:
      summary: Get all products (v2)
      description: Enhanced product listing with advanced filtering, sorting, and analytics
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/InStockFilter'
        - $ref: '#/components/parameters/MinPriceFilter'
        - $ref: '#/components/parameters/MaxPriceFilter'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/IncludeAnalyticsParam'
      responses:
        '200':
          description: Enhanced product list with metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new product (v2)
      description: Create a new product with enhanced features and validation
      tags:
        - Products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/products/{product_id}:
    get:
      summary: Get product by ID (v2)
      description: Retrieve a specific product with enhanced data and recommendations
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/IncludeRecommendationsParam'
        - $ref: '#/components/parameters/IncludeAnalyticsParam'
      responses:
        '200':
          description: Enhanced product data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Search Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    ProductIdParam:
      name: product_id
      in: path
      required: true
      description: Unique identifier for the product
      schema:
        type: string
        example: "product-123"
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    CategoryFilter:
      name: category
      in: query
      description: Filter by product category
      schema:
        type: string
        example: "Electronics"
    SearchQuery:
      name: search
      in: query
      description: Search products by name or description
      schema:
        type: string
        example: "wireless"
    InStockFilter:
      name: in_stock
      in: query
      description: Filter by stock availability
      schema:
        type: boolean
        example: true
    MinPriceFilter:
      name: min_price
      in: query
      description: Minimum price filter
      schema:
        type: number
        format: float
        minimum: 0
        example: 10.00
    MaxPriceFilter:
      name: max_price
      in: query
      description: Maximum price filter
      schema:
        type: number
        format: float
        minimum: 0
        example: 500.00
    SortByParam:
      name: sort_by
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [name, price, created_at, updated_at, rating]
        default: created_at
        example: price
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
        example: asc
    IncludeRecommendationsParam:
      name: include_recommendations
      in: query
      description: Include product recommendations
      schema:
        type: boolean
        default: false
        example: true
    IncludeAnalyticsParam:
      name: include_analytics
      in: query
      description: Include analytics data
      schema:
        type: boolean
        default: false
        example: true
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
          description: Unique product identifier
          example: "product-123"
        name:
          type: string
          description: Product name
          example: "Wireless Bluetooth Headphones"
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
          example: 79.99
        category:
          type: string
          description: Product category
          example: "Electronics"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        in_stock:
          type: boolean
          description: Whether the product is in stock
          example: true
        stock_quantity:
          type: integer
          description: Current stock quantity
          minimum: 0
          example: 50
        created_at:
          type: string
          format: date-time
          description: Product creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Product last update timestamp
        version:
          type: string
          description: API version that created this response
          example: "1.0.0"
    EnhancedProduct:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
                format: uri
                example: "https://cdn.retailgenie.com/products/123/image1.jpg"
            rating:
              type: number
              format: float
              minimum: 0
              maximum: 5
              example: 4.5
            review_count:
              type: integer
              minimum: 0
              example: 127
            tags:
              type: array
              items:
                type: string
              example: ["wireless", "bluetooth", "noise-cancelling"]
            dimensions:
              type: object
              properties:
                length:
                  type: number
                  format: float
                width:
                  type: number
                  format: float
                height:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
            manufacturer:
              type: string
              example: "TechCorp"
            warranty_months:
              type: integer
              minimum: 0
              example: 24
            recommendations:
              type: array
              items:
                $ref: '#/components/schemas/Product'
            analytics:
              $ref: '#/components/schemas/ProductAnalytics'
    ProductCreateRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          description: Product name
          example: "Wireless Bluetooth Headphones"
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
          example: 79.99
        category:
          type: string
          description: Product category
          example: "Electronics"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones"
        in_stock:
          type: boolean
          description: Whether the product is in stock
          default: true
        stock_quantity:
          type: integer
          description: Initial stock quantity
          minimum: 0
          default: 0
    EnhancedProductCreateRequest:
      allOf:
        - $ref: '#/components/schemas/ProductCreateRequest'
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
                format: uri
            tags:
              type: array
              items:
                type: string
            dimensions:
              type: object
              properties:
                length:
                  type: number
                  format: float
                width:
                  type: number
                  format: float
                height:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
            manufacturer:
              type: string
            warranty_months:
              type: integer
              minimum: 0
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Product name
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
        category:
          type: string
          description: Product category
        description:
          type: string
          description: Product description
        in_stock:
          type: boolean
          description: Whether the product is in stock
        stock_quantity:
          type: integer
          description: Stock quantity
          minimum: 0
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        count:
          type: integer
          description: Total number of products returned
          example: 25
        total:
          type: integer
          description: Total number of products matching criteria
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pages:
          type: integer
          description: Total number of pages
          example: 8
        version:
          type: string
          example: "1.0.0"
    EnhancedProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/EnhancedProduct'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        filters:
          $ref: '#/components/schemas/AppliedFilters'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'

    # Search Schemas
...
    ProductAnalytics:
      type: object
      properties:
        views:
          type: integer
          example: 1247
        conversions:
          type: integer
          example: 23
        conversion_rate:
          type: number
          format: float
          example: 0.018
        revenue:
          type: number
          format: float
          example: 1839.77
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    AppliedFilters:
      type: object
      properties:
        category:
          type: string
        price_range:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
        in_stock:
          type: boolean
        search_query:
          type: string
    ResponseMetadata:
      type: object
      properties:
        version:
          type: string
          example: "2.1.0"
        timestamp:
          type: string
          format: date-time
        execution_time_ms:
          type: number
          format: float
          example: 45.2
        rate_limit:
          type: object
          properties:
            remaining:
              type: integer
              example: 995
            reset_time:
              type: string
              format: date-time

    # Error Schemas
    Error:
      type: object
      required:
        - error
        - status_code
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        status_code:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          example: "2023-01-15T10:30:00Z"
        path:
          type: string
          description: API endpoint that generated the error
          example: "/api/products/invalid-id"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        request_id:
          type: string
          description: Unique request identifier for debugging
          example: "req-123-abc"
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            status_code: 400
            timestamp: "2023-01-15T10:30:00Z"
            path: "/api/products"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            status_code: 401
            timestamp: "2023-01-15T10:30:00Z"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            status_code: 404
            timestamp: "2023-01-15T10:30:00Z"
            path: "/api/products/invalid-id"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            status_code: 500
            timestamp: "2023-01-15T10:30:00Z"

# Example webhook configurations for real-time updates
...